Move HttpServer::HTML into its own file
